Add testing for caching jobs.
This removes Resource caching because it wasn't paging resources and I
don't see a place in the code where having a local copy of all the
resources is actually necessary.

We may need this when we validate which AbID to put in next, but if so
having a local copy of every resource will be too intensive anyways.