Rewrite one method in rust
With much thanks to this blog post:
https://blog.codeminer42.com/integrating-ruby-with-rust-with-ffi/

In order to deploy this, we would first need to add cargo to the
lib-jobs boxes (perhaps through apt).

Currently, the code is in a crate of type cdylib, which does not
support doctests! 😢 I wonder if we could use the cdylib as a
compatibility layer, and then use a lib crate (with doctests)
to actually provide the functionality.

We should also make sure we are running the following regularly,
perhaps as part of CI?
 * `cargo clippy`
 * `cargo test`
 * `cargo fmt`
 * `cargo doc`

I am also confused as to whether or not this contains a memory
leak.  FFI is returning a two-element array, with our string and
then a pointer.  But if I try to free the pointer afterwards
(which I could be doing wrong), I get an error that it hasn't
been allocated?  Very mysterious (and ironic that we have to
worry about it when using memory safe languages like Ruby and Rust).